back/requirements.txt
Flask>=2.3
mysql-connector-python>=8.0
pytest>=7.0
requests
python-dotenv
